# swagger-response.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Parroquia API",
    "version": "1.0.0",
    "description": "Sistema de gestión parroquial - API REST completa con autenticación y gestión de usuarios",
    "contact": {
      "name": "API Support",
      "email": "support@parroquia.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "http://206.62.139.100:3000/",
      "description": "Test server"
    },
    {
      "url": "https://api.parroquia.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "ID único del usuario (UUID)",
            "example": "2acc2c01-91d8-4d3a-a184-f002304620ec"
          },
          "correo_electronico": {
            "type": "string",
            "format": "email",
            "description": "Correo electrónico del usuario",
            "example": "usuario@example.com"
          },
          "primer_nombre": {
            "type": "string",
            "description": "Primer nombre del usuario",
            "example": "Juan"
          },
          "segundo_nombre": {
            "type": "string",
            "nullable": true,
            "description": "Segundo nombre del usuario (opcional)",
            "example": "Carlos"
          },
          "primer_apellido": {
            "type": "string",
            "description": "Primer apellido del usuario",
            "example": "Pérez"
          },
          "segundo_apellido": {
            "type": "string",
            "nullable": true,
            "description": "Segundo apellido del usuario (opcional)",
            "example": "García"
          },
          "activo": {
            "type": "boolean",
            "description": "Estado activo del usuario",
            "example": true
          },
          "email_verificado": {
            "type": "boolean",
            "description": "Si el email ha sido verificado",
            "example": true
          },
          "fecha_verificacion_email": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Fecha de verificación del email",
            "example": "2025-07-31T10:30:00.000Z"
          },
          "expira_token_reset": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Fecha de expiración del token de reset",
            "example": "2025-07-31T12:30:00.000Z"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Lista de roles asignados al usuario",
            "example": [
              "Administrador"
            ]
          }
        }
      },
...
      "UserUpdate": {
        "type": "object",
        "properties": {
          "primer_nombre": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "description": "Primer nombre del usuario (opcional para actualización)",
            "example": "Juan",
            "pattern": "^[a-zA-ZÀ-ÿ\\u00f1\\u00d1\\s]+$"
          },
          "segundo_nombre": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "nullable": true,
            "description": "Segundo nombre del usuario (opcional)",
            "example": "Carlos",
            "pattern": "^[a-zA-ZÀ-ÿ\\u00f1\\u00d1\\s]+$"
          },
          "primer_apellido": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "description": "Primer apellido del usuario (opcional para actualización)",
            "example": "Pérez",
            "pattern": "^[a-zA-ZÀ-ÿ\\u00f1\\u00d1\\s]+$"
          },
          "segundo_apellido": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "nullable": true,
            "description": "Segundo apellido del usuario (opcional)",
            "example": "García",
            "pattern": "^[a-zA-ZÀ-ÿ\\u00f1\\u00d1\\s]+$"
          },
          "correo_electronico": {
            "type": "string",
            "format": "email",
            "maxLength": 100,
            "description": "Correo electrónico único (opcional para actualización)",
            "example": "juan.perez@yopmail.com"
          },
          "telefono": {
            "type": "string",
            "minLength": 10,
            "maxLength": 20,
            "description": "Número de teléfono del usuario (opcional)",
            "example": "+57 300 123 4567",
            "pattern": "^[\\+]?[0-9\\s\\-\\(\\)]+$"
          }
        }
      },
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "error"
            ],
            "description": "Estado de la respuesta",
            "example": "success"
          },
          "message": {
            "type": "string",
            "description": "Mensaje descriptivo",
            "example": "Operación realizada exitosamente"
          },
          "data": {
            "type": "object",
            "description": "Datos de respuesta (opcional)"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "error"
            ],
            "example": "error"
          },
          "message": {
            "type": "string",
            "description": "Mensaje de error",
            "example": "Ha ocurrido un error"
          },
          "code": {
            "type": "string",
            "description": "Código de error específico",
            "example": "VALIDATION_ERROR"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "description": "Campo que causó el error",
                  "example": "email"
                },
                "message": {
                  "type": "string",
                  "description": "Mensaje específico del error",
                  "example": "El formato del email no es válido"
                }
              }
            },
            "description": "Detalles específicos de errores de validación"
          }
        }
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Token de acceso requerido o inválido",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": "error",
              "message": "Token de acceso requerido",
              "code": "UNAUTHORIZED"
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Acceso denegado - permisos insuficientes",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": "error",
              "message": "Acceso denegado",
              "code": "FORBIDDEN"
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Recurso no encontrado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": "error",
              "message": "Usuario no encontrado",
              "code": "NOT_FOUND"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Error de validación en los datos enviados",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": "error",
              "message": "Errores de validación",
              "code": "VALIDATION_ERROR",
              "errors": [
                {
                  "field": "email",
                  "message": "El formato del email no es válido"
                },
                {
                  "field": "password",
                  "message": "La contraseña debe tener al menos 8 caracteres"
                }
              ]
            }
          }
        }
      },
      "ConflictError": {
        "description": "Conflicto - el recurso ya existe o hay un conflicto con el estado actual",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": "error",
              "message": "El email ya está registrado",
              "code": "CONFLICT"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Error interno del servidor",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": "error",
              "message": "Error interno del servidor",
              "code": "INTERNAL_SERVER_ERROR"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/users": {
      "get": {
        "summary": "Obtener todos los usuarios activos",
        "description": "Recupera una lista de todos los usuarios activos del sistema (solo administradores)",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuarios obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "users": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/User"
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/users/deleted": {
      "get": {
        "summary": "Obtener usuarios eliminados",
        "description": "Recupera una lista de todos los usuarios eliminados (soft delete) del sistema (solo administradores)",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuarios eliminados obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "users": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/User"
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "summary": "Obtener usuario por ID",
        "description": "Recupera un usuario específico por su ID (administrador o propietario)",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID único del usuario (UUID)",
            "example": "31cb8fe4-bb24-4ad8-af66-d8c900d13c2a"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario obtenido exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "user": {
                              "$ref": "#/components/schemas/User"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "summary": "Actualizar usuario",
        "description": "Actualiza la información de un usuario existente (administrador o propietario)",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID único del usuario (UUID)",
            "example": "31cb8fe4-bb24-4ad8-af66-d8c900d13c2a"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              },
              "example": {
                "firstName": "Juan Carlos",
                "lastName": "Pérez García",
                "email": "juan.carlos@yopmail.com",
                "role": "surveyor"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "user": {
                              "$ref": "#/components/schemas/User"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "409": {
            "$ref": "#/components/responses/ConflictError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "summary": "Eliminar usuario",
        "description": "Elimina un usuario del sistema (soft delete - solo administradores)",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID único del usuario (UUID)",
            "example": "31cb8fe4-bb24-4ad8-af66-d8c900d13c2a"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario eliminado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  }
}